Backed out changeset: 7e0aaedeea4f

+++ KV_Team_OSD/Screen.h
@@ -6,7 +6,7 @@
 	void displayMode(void);
 	void displayArmed(void);
 	void displayCallsign(void);
-	void displayHorizon(int16_t rollAngle, int16_t pitchAngle);
+	void displayHorizon(int rollAngle, int pitchAngle);
 	void displayVoltage(void);
 	void displayVidVoltage(void);
 	void displayCurrentThrottle(void);
